Require hook in standalone mode and add JS target
Koja is meant to be used in conjunction with kmcp. When used as a standalone library, users will need to invoke an initialization method at application startup to push the generated KDocs into the global registry. This is necessary for the IR plugin to find a hook. This is not necessary when used with kmcp as the entry points there are known.

Enforce top-level and refactor